+++ src/CoreShop/Bundle/CoreBundle/CoreExtension/StorePrice.php
@@ -229,7 +229,13 @@ class StorePrice extends Model\DataObject\ClassDefinition\Data
      */
     public function preGetData($object, $params = [])
     {
-        $data = $object->{$this->getName()};
+        //TODO: Remove once CoreShop requires min Pimcore 5.5
+        if (method_exists($object, 'getObjectVar')) {
+            $data = $object->getObjectVar($this->getName());
+        } else {
+            $data = $object->{$this->getName()};
+        }
+
         if (!in_array($this->getName(), $object->getO__loadedLazyFields())) {
             $data = $this->load($object, ['force' => true]);